Tweaks to support WASM downloading on-demand instead of preinstalling from NPM
