+++ package.json
@@ -34,70 +34,8 @@
         "jsel": "^1.1.6",
         "minimist": "^1.2.5",
         "semver": "^7.3.5",
-        "solc-0.4.13": "npm:solc@0.4.13",
-        "solc-0.4.14": "npm:solc@0.4.14",
-        "solc-0.4.15": "npm:solc@0.4.15",
-        "solc-0.4.16": "npm:solc@0.4.16",
-        "solc-0.4.17": "npm:solc@0.4.17",
-        "solc-0.4.18": "npm:solc@0.4.18",
-        "solc-0.4.19": "npm:solc@0.4.19",
-        "solc-0.4.20": "npm:solc@0.4.20",
-        "solc-0.4.21": "npm:solc@0.4.21",
-        "solc-0.4.22": "npm:solc@0.4.22",
-        "solc-0.4.23": "npm:solc@0.4.23",
-        "solc-0.4.24": "npm:solc@0.4.24",
-        "solc-0.4.25": "npm:solc@0.4.25",
         "solc-0.4.26": "npm:solc@0.4.26",
-        "solc-0.5.0": "npm:solc@0.5.0",
-        "solc-0.5.1": "npm:solc@0.5.1",
-        "solc-0.5.10": "npm:solc@0.5.10",
-        "solc-0.5.11": "npm:solc@0.5.11",
-        "solc-0.5.12": "npm:solc@0.5.12",
-        "solc-0.5.13": "npm:solc@0.5.13",
-        "solc-0.5.14": "npm:solc@0.5.14",
-        "solc-0.5.15": "npm:solc@0.5.15",
-        "solc-0.5.16": "npm:solc@0.5.16",
-        "solc-0.5.17": "npm:solc@0.5.17",
-        "solc-0.5.2": "npm:solc@0.5.2",
-        "solc-0.5.3": "npm:solc@0.5.3",
-        "solc-0.5.4": "npm:solc@0.5.4",
-        "solc-0.5.5": "npm:solc@0.5.5",
-        "solc-0.5.6": "npm:solc@0.5.6",
-        "solc-0.5.7": "npm:solc@0.5.7",
-        "solc-0.5.8": "npm:solc@0.5.8",
-        "solc-0.5.9": "npm:solc@0.5.9",
-        "solc-0.6.0": "npm:solc@0.6.0",
-        "solc-0.6.1": "npm:solc@0.6.1",
-        "solc-0.6.10": "npm:solc@0.6.10",
-        "solc-0.6.11": "npm:solc@0.6.11",
-        "solc-0.6.12": "npm:solc@0.6.12",
-        "solc-0.6.2": "npm:solc@0.6.2",
-        "solc-0.6.3": "npm:solc@0.6.3",
-        "solc-0.6.4": "npm:solc@0.6.4",
-        "solc-0.6.5": "npm:solc@0.6.5",
-        "solc-0.6.6": "npm:solc@0.6.6",
-        "solc-0.6.7": "npm:solc@0.6.7",
-        "solc-0.6.8": "npm:solc@0.6.8",
-        "solc-0.6.9": "npm:solc@0.6.9",
-        "solc-0.7.0": "npm:solc@0.7.0",
-        "solc-0.7.1": "npm:solc@0.7.1",
-        "solc-0.7.2": "npm:solc@0.7.2",
-        "solc-0.7.3": "npm:solc@0.7.3",
-        "solc-0.7.4": "npm:solc@0.7.4",
-        "solc-0.7.5": "npm:solc@0.7.5",
-        "solc-0.7.6": "npm:solc@0.7.6",
-        "solc-0.8.0": "npm:solc@0.8.0",
-        "solc-0.8.1": "npm:solc-0.8.1-fixed@0.8.1",
-        "solc-0.8.10": "npm:solc@0.8.10",
         "solc-0.8.11": "npm:solc@0.8.11",
-        "solc-0.8.2": "npm:solc@0.8.2",
-        "solc-0.8.3": "npm:solc@0.8.3",
-        "solc-0.8.4": "npm:solc@0.8.4",
-        "solc-0.8.5": "npm:solc@0.8.5",
-        "solc-0.8.6": "npm:solc@0.8.6",
-        "solc-0.8.7": "npm:solc@0.8.7-fixed",
-        "solc-0.8.8": "npm:solc@0.8.8",
-        "solc-0.8.9": "npm:solc@0.8.9",
         "src-location": "^1.1.0",
         "web3-eth-abi": "^1.7.0"
     },
@@ -114,7 +52,7 @@
         "eslint-config-prettier": "^8.3.0",
         "eslint-plugin-prettier": "^4.0.0",
         "expect": "^27.4.6",
-        "mocha": "^9.1.4",
+        "mocha": "^9.2.0",
         "nyc": "^15.1.0",
         "peggy": "^1.2.0",
         "prettier": "2.5.1",
