+++ src/compile/kinds/index.ts
@@ -1,4 +1,2 @@
 export * from "./compiler";
-export * from "./native";
-export * from "./wasm";
 export * from "./md";
